core: move more storage functions to storage.c
extstore.h is now only used from storage.c. starting a path towards
getting the storage interface to be more generalized.

should be no functional changes.
